my-iot-core-app/requirements.txt
boto3
paho-mqtt
jsonschema